move insights to component
